+++ analysis/backend/v2-flat/genProperties.ts
@@ -2,46 +2,14 @@ import * as fs from "fs";
 import path from "path";
 import {BACKEND_DIR, IMAGES_DIR, README_PATH} from "../const";
 import {Property} from "./ds";
-import {read} from "fs";
-
-let provinceNamesUnderImagesDir = fs.readdirSync(IMAGES_DIR, {withFileTypes: true})
-    .filter(d => d.isDirectory())
-    .map(d => d.name)
-
-function getProvinceNameUnderImagesDir(inputProvinceName: string): string {
-    for (let provinceNameUnderImagesDir of provinceNamesUnderImagesDir) {
-        if (inputProvinceName.includes(provinceNameUnderImagesDir))
-            return provinceNameUnderImagesDir
-    }
-    console.warn(`province dir of ${inputProvinceName} not exists, creating...`)
-    fs.mkdirSync(path.join(IMAGES_DIR, inputProvinceName))
-    return inputProvinceName
-}
-
-function getCityNameUnderProvinceImagesDir(provinceDir: string, inputCityName: string): string | undefined {
-    for (let cityNameUnderProvinceImagesDir of fs.readdirSync(path.join(IMAGES_DIR, provinceDir))) {
-        if (inputCityName.includes(cityNameUnderProvinceImagesDir))
-            return cityNameUnderProvinceImagesDir
-    }
-    // console.error(`not find a proper city name under province images dir ${provinceDir} for ${inputCityName}`)
-}
-
-function getImageUriRobust(provinceDir: string, cityDir: string | undefined, fileName: string): string | undefined {
-    let imgUri = path.join(provinceDir, fileName)
-    if (fs.existsSync(path.join(IMAGES_DIR, imgUri)))
-        return imgUri
-    if (cityDir) {
-        imgUri = path.join(provinceDir, cityDir, fileName)
-        if (!fs.existsSync(path.join(IMAGES_DIR, imgUri)))
-            console.error(`not found image uri of ${imgUri}`)
-        return imgUri
-    }
-}
+import {getCityNameUnderProvinceImagesDir, getImageUriRobust, getProvinceNameUnderImagesDir} from "./utils";
 
 function analyze() {
     let startCollecting = false
-    let isModified = false
-    let newReadme = ""
+
+    // 用于省份排序输出
+    let headingPart = ''
+    let provinceDict = new Map<string, string>()
 
     // 收集楼盘
     let collectedProperties: Property[] = []
@@ -93,67 +61,69 @@ function analyze() {
         // 等待匹配到第一个省份
         if (matchProvince)
             startCollecting = true
-        if (startCollecting) {
 
-            // 解析省份
-            if (matchProvince) {
-                countProvinces += 1
-                if (curProvince) validateProvince() // 后一个省份解析前一个省份
-
-                curProvince = matchProvince[1]
-                provinceDir = getProvinceNameUnderImagesDir(curProvince)
-
-                markedTotalInProvince = parseInt(matchProvince[2])
-                console.log(`parsing province ${curProvince}`)
-                cumTotalInProvince = 0
-            }
-
-            // 解析城市
-            else if (matchCity) {
-                countCities += 1
-                curCity = matchCity[1]
-                cityDir = getCityNameUnderProvinceImagesDir(provinceDir, curCity)
-                markedTotalInCity = parseInt(matchCity[2])
-                let propertiesInCity = matchCity[3].split("，")
-                // 验证城市统计
-                if (propertiesInCity.length !== markedTotalInCity)
-                    console.error(`    properties count in city ${curCity} failed to validate, calculated: ${String(propertiesInCity.length).padStart(3)}, marked: ${String(markedTotalInCity).padStart(3)}`)
-                cumTotalInProvince += propertiesInCity.length
-
-                propertiesInCity.forEach((propertyStr: string) => {
-                    let property: Property = {name: "", city: curCity, province: curProvince}
-                    const matchPropertyLink = propertyStr.match(/\s*\[(.*?)\]\((.*?)\)/)
-                    if (matchPropertyLink) {
-                        property.name = matchPropertyLink[1]
-
-                        // 重新归档到省份文件夹内
-                        let parsedLink = matchPropertyLink[2]
-                        let parsedLinkFrags = parsedLink.split('/')
-
-                        let fileName = parsedLinkFrags[parsedLinkFrags.length - 1]
-                            || property.name // 有时候链接只有一个光秃秃的省份，所以用名字去匹配
-                        let imageUri = getImageUriRobust(provinceDir, cityDir, fileName)
-                        let newLink = [path.basename(IMAGES_DIR), imageUri].join('/')
-                        property.link = newLink
-                        if (parsedLink !== newLink) {
-                            isModified = true
-                            line = line.replace(parsedLink, newLink)
-                        }
-                    } else {
-                        property.name = propertyStr.split(" ").join("")
-                    }
-                    const matchPropertyMonth = property.name.match(/(.*?)\s*（(\d+)月）/)
-                    if (matchPropertyMonth) {
-                        property.name = matchPropertyMonth[1]
-                        property.month = parseInt(matchPropertyMonth[2])
-                    }
-
-                    collectedProperties.push(property)
-                })
-            }
+        if (!startCollecting) {
+            headingPart += line + "\n"
+            return
         }
 
-        newReadme += line + '\n'
+        // 解析省份
+        if (matchProvince) {
+            countProvinces += 1
+            if (curProvince) validateProvince() // 后一个省份解析前一个省份
+
+            curProvince = matchProvince[1]
+            provinceDir = getProvinceNameUnderImagesDir(curProvince)
+            provinceDict[curProvince] = ""
+
+            markedTotalInProvince = parseInt(matchProvince[2])
+            console.log(`parsing province ${curProvince}`)
+            cumTotalInProvince = 0
+        }
+
+        // 解析城市
+        else if (matchCity) {
+            countCities += 1
+            curCity = matchCity[1]
+            cityDir = getCityNameUnderProvinceImagesDir(provinceDir, curCity)
+            markedTotalInCity = parseInt(matchCity[2])
+            let propertiesInCity = matchCity[3].split("，")
+            // 验证城市统计
+            if (propertiesInCity.length !== markedTotalInCity)
+                console.error(`    properties count in city ${curCity} failed to validate, calculated: ${String(propertiesInCity.length).padStart(3)}, marked: ${String(markedTotalInCity).padStart(3)}`)
+            cumTotalInProvince += propertiesInCity.length
+
+            propertiesInCity.forEach((propertyStr: string) => {
+                let property: Property = {name: "", city: curCity, province: curProvince}
+                const matchPropertyLink = propertyStr.match(/\s*\[(.*?)\]\((.*?)\)/)
+                if (matchPropertyLink) {
+                    property.name = matchPropertyLink[1]
+
+                    // 重新归档到省份文件夹内
+                    let parsedLink = matchPropertyLink[2]
+                    let parsedLinkFrags = parsedLink.split('/')
+
+                    let fileName = parsedLinkFrags[parsedLinkFrags.length - 1]
+                        || property.name // 有时候链接只有一个光秃秃的省份，所以用名字去匹配
+                    let imageUri = getImageUriRobust(provinceDir, cityDir, fileName)
+                    let newLink = [path.basename(IMAGES_DIR), imageUri].join('/')
+                    property.link = newLink
+                    if (parsedLink !== newLink)
+                        line = line.replace(parsedLink, newLink)
+                } else {
+                    property.name = propertyStr.split(" ").join("")
+                }
+                const matchPropertyMonth = property.name.match(/(.*?)\s*（(\d+)月）/)
+                if (matchPropertyMonth) {
+                    property.name = matchPropertyMonth[1]
+                    property.month = parseInt(matchPropertyMonth[2])
+                }
+
+                collectedProperties.push(property)
+            })
+        }
+
+        provinceDict[curProvince] += line + '\n'
     }
 
     // parse
@@ -166,13 +136,15 @@ function analyze() {
     fs.writeFileSync(path.join(__dirname, "properties.json"), data, "utf-8")
 
     // rewrite
-    if (isModified) {
-        const readmePathBackedUp = path.join(BACKEND_DIR, "tmp/README.md")
-        console.log('backup README.md to ' + readmePathBackedUp)
-        fs.cpSync(README_PATH, readmePathBackedUp)
-        console.log('rewriting README.md')
-        fs.writeFileSync(README_PATH, newReadme, 'utf-8')
-    }
+    const readmePathBackedUp = path.join(BACKEND_DIR, "tmp/README.md")
+    console.log('backup README.md to ' + readmePathBackedUp)
+    fs.cpSync(README_PATH, readmePathBackedUp)
+    // 按拼音排序，ref: https://blog.csdn.net/qq_27674439/article/details/115406758
+    console.log("sorting provinces part based on pronunciation")
+    let provinceKeys = Object.keys(provinceDict).sort((word1, word2) => word1.localeCompare(word2, 'zh'))
+    let provincesPart = provinceKeys.map(k => provinceDict[k]).join('\n')
+    console.log('rewriting README.md')
+    fs.writeFileSync(README_PATH, [headingPart, provincesPart].join('\n').replace(/\n{2,9}/g, '\n\n'), 'utf-8')
 }
 
 analyze()
